refector: abstract form-rendering logic into a separate component
